refactor MigrationFile namespace
